clean other table

+++ antarest/output/storage/v2/storage.py
@@ -321,11 +321,14 @@ class V2OutputStorage(IOutputStorage):
 
     @override
     def delete_outputs(self, study_id: str) -> None:
-        for output_metadata in self._repository.search_output_metadata(study_id):
+        outputs = self._repository.search_output_metadata(study_id)
+        for output_metadata in outputs:
             output_id = output_metadata.output_name
             self._archive_storage.delete_file(_archive_id(study_id, output_id))
             shutil.rmtree(parquet_output_dir(self._variables_dir, study_id, output_id), ignore_errors=True)
-        self._repository.delete_outputs(study_id)
+
+        if outputs:
+            self._repository.delete_outputs(study_id)  # Needed as we don't have a ForeignKey to the `Study` table
 
     @override
     def export_output(self, study_id: str, output_id: str, target: Path) -> None:
